Improve CLI
Added actions:
RemoveAndDelete
Delete
Restore
List2
PopulateConnectionID
